add sign-up-username
