+++ my-app-ts/src/notespace
@@ -1,13 +1,13 @@
-import { useEffect, useState } from "react";
+/* import { useEffect, useState } from "react";
 import './App.css';
-import Form from './pages/Form';
-import UsersDisplay from "./pages/usersDisplay";
+import Form from './components/Form';
+import UsersDisplay from "./components/usersDisplay";
 import { fetchUsers } from "./utils/FetchUsers";
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Register from "./pages/signUp/Register";
-import Login from "./pages/Login";
-import Mypage from "./pages/myPage/Mypage";
+import Register from "./components/Register";
+import Login from "./components/Login";
+import Mypage from "./components/Mypage";
 
 function App() {
   //userの型定義
@@ -20,6 +20,8 @@ function App() {
   const initialURL = "http://localhost:8080/user"; //endpoint
   const [loading, setLoading] = useState<Boolean>(true); //最初にloading出したいのでtrue
   const [usersData, setUsersData] = useState<User[]>([]);//userの全データをstateに設定
+
+  console.log(typeof(usersData))
 
   //reactアプリ立ち上げ時にuserデータを取得
   useEffect(() => {
@@ -81,6 +83,7 @@ function App() {
   return (
     <div className='App'>
         <header className='App-header'>
+          <h2 className='header-string'>User Register</h2>
         </header>
         <div className="container">
           <BrowserRouter>
@@ -91,8 +94,8 @@ function App() {
             </Routes>
           </BrowserRouter>
         </div>
-        {/* <Form onSubmit={onSubmit}/> */}
-        {/* <div className='users-display'>
+        <Form onSubmit={onSubmit}/>
+        <div className='users-display'>
           {loading ? (
               <p className="loader">ロード中。。。</p>
           ) : (
@@ -102,9 +105,10 @@ function App() {
                 })}
               </div>
           )}
-        </div> */}
+        </div>
       </div>
   );
 }
 
 export default App;
+ */
