+++ my-app-ts/src/pages/myPage/Mypage.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { onAuthStateChanged, signOut } from "firebase/auth";
-import { auth } from "../firebase";
+import { auth } from "../../firebase";
 import {
   useNavigate,
   Navigate
